Pipeline resiliency + hardport fixes
Pipeline:
- Architect stage: cache-load architecture.json when present, fall back
  to LLM only if parse fails. Prior behavior was always-regenerate, which
  turned a transient LLM hiccup during architect into a whole-run abort.
- Trait extractor: dedupe by name before appending. With cache-load in
  place, post-architect trait extraction runs against an already-enriched
  arch; without dedupe it appends duplicates on every run, producing
  "Compressor defined multiple times" compile errors in lib.rs.

Skeleton:
- Auto-import types from dep crates: when a dep crate is the shared-types
  crate (name ends in "-types"), emit `pub use <crate>::*` instead of
  only importing named error types. Fixes "cannot find type TreeElement"
  when traits reference non-error struct types from zlib-types.

Hardports:
- detect_data_type: return i32 instead of u8 to match C's `int` signature
  and the observer binding's return_rust_type.
- adler32_z: explicit `NMAX as usize` cast. Constants extractor types
  bare-decimal #defines as i32 (correct), but the usize arithmetic in
  the loop needs a cast.

C-shim observer binding:
- CShimField: add `rust_field` + `rust_write_template` for shim-vs-Rust
  field-shape translation. detect_data_type's shim sees `dyn_ltree_freq:
  Vec<u16>` while Rust sees `dyn_ltree: Vec<(u16,u16)>`; template bridges.
- fuzz_observer_shim: emit fully-rendered statements under `__stmt__<N>`
  keys so the test emitter doesn't have to know the field mapping.
- Observer dispatch path in tdd_generator remains disabled until the next
  run verifies the baseline is stable; infrastructure is ready to re-enable.

+++ alchemist/references/impls/zlib_hardports/zlib-trees/trees/detect_data_type.rs
@@ -1,19 +1,19 @@
-pub fn detect_data_type(s: &DeflateState) -> u8 {
-    // Port of trees.c:detect_data_type.
-    // Classify the pending block as binary (0), text (1), or unknown (2).
+pub fn detect_data_type(s: &DeflateState) -> i32 {
+    // Port of trees.c:detect_data_type — C signature returns `int`, so
+    // the Rust return type is i32 to match byte-for-byte across the FFI.
+    // Classify the pending block as BINARY (0), TEXT (1), or UNKNOWN (2).
     // Algorithm (RFC 1950 is silent; zlib's heuristic):
     //   - If any frequency for a non-text control byte (0-31 except 9,10,13)
     //     is non-zero → BINARY.
     //   - Else, if any text byte (9,10,13 or 32+) has frequency → TEXT.
     //   - Else → UNKNOWN.
     //
-    // In the Rust generated types, dyn_ltree is Vec<(u16, u16)> = (freq, len).
+    // Rust's DeflateState exposes dyn_ltree as Vec<(u16, u16)> = (freq, len).
     // Walk [0..31] for binary control chars (mask 0xF3FFC07F per zlib),
     // then [32..LITERALS=256] for text.
 
     // Block-specific bitmask: each bit N means "byte N counts as BINARY".
-    // The mask 0xF3FFC07F = 1111_0011_1111_1111_1100_0000_0111_1111 in
-    // bits 0..31 — all control chars EXCEPT 9 (tab), 10 (LF), 13 (CR).
+    // 0xF3FFC07F = all control chars EXCEPT 9 (tab), 10 (LF), 13 (CR).
     let block_mask: u32 = 0xF3FF_C07F;
     let mut mask = block_mask;
     let mut n: usize = 0;
